added TRPC (#45)
* re-wrote population script (again)

* removed duplicate components

* new population script

* generated Zod models

* finished seed script

* removed dev api endpoints and dev-panel page, moved all functionality to seed and unseed scripts, removed replaced data folder with db folder

* started writing database procedures

* updated postcss version

* started fixing dark mode inconsistencies

* added trpc to project

* begun work on project procedures

* removed old procedures folder

* finished first version of projectRouter

* finished first version of studentRouter

+++ src/types/zod/preference.ts
@@ -4,7 +4,7 @@ import { CompleteProject, RelatedProjectModel, CompleteStudent, RelatedStudentMo
 export const PreferenceModel = z.object({
   projectId: z.string(),
   studentId: z.string(),
-  rank: z.number().int(),
+  rank: z.number(),
 })
 
 export interface CompletePreference extends z.infer<typeof PreferenceModel> {
